+++ lizard_rainapp/urls.py
@@ -9,24 +9,25 @@ from django.http import HttpResponseServerError
 from django.template import Context
 from django.template import loader
 
+from lizard_fewsjdbc.views import JdbcSourceView, HomepageView
+
 admin.autodiscover()
 handler404  # pyflakes
 
 urlpatterns = patterns(
     '',
     url(r'^$',
-        'lizard_fewsjdbc.views.homepage',
-        {'template': 'lizard_rainapp/homepage.html'},
+        HomepageView.as_view(template_name='lizard_rainapp/homepage.html'),
         name="lizard_rainapp.homepage",
         ),
     url(r'^fews_jdbc/(?P<jdbc_source_slug>.*)/$',
-        'lizard_fewsjdbc.views.jdbc_source',
-        {'adapter_class': 'adapter_rainapp',
-         'filter_url_name': 'lizard_rainapp.jdbc_source'},
+        JdbcSourceView.as_view(adapter_class='adapter_rainapp',
+                               filter_url_name="lizard_rainapp.jdbc_source"),
         name="lizard_rainapp.jdbc_source",
         ),
     (r'^admin/', include(admin.site.urls)),
     )
+
 
 
 if settings.DEBUG: